requirements.txt
PyYAML
dpkt